FlagForm will now parse the Recon options bitmask when opened.
The Recon must be connected and the switch turned to I/O. With both reading
and writing the flag bitmask, it is NOT necessary to have clicked the Connect
button on the main screen--the form initiates its own connection.

These values will populate the combo boxes on the form. A defaults button has
been added to the form which will reset all values to US. The user must always
click apply and wait for confirmation in order to actually write the settings.