More improvements to file naming process.
Now when a Recon is connected that has data sessions, the name field
will be populated with the default and selected so that the user knows
it's editable. If the checkbox is selected, text will be pulled from
the first line of Test Site Info.

Error checking is in place so that if the checkbox is selected but
Test Site Info is empty, or if the name field is empty, the default
value will still be used, so we should always get a file.